chore(ci): enable CircleCI snapshot and doc publishing (#11608)
<!-- 
Filling out this template is required! Do not delete it when submitting a Pull Request! Without this information, your Pull Request may be auto-closed.
-->
## PR Checklist
Please check that your PR fulfills the following requirements:
- [x] The commit message follows [our guidelines](https://github.com/angular/material/blob/master/.github/CONTRIBUTING.md#-commit-message-format)
- [x] Tests for the changes have been added or this is not a bug fix / enhancement
- [x] Docs have been added, updated, or were not required

## PR Type
What kind of change does this PR introduce?
<!-- Please check the one that applies to this PR using "x". -->
```
[ ] Bugfix
[ ] Enhancement
[ ] Documentation content changes
[ ] Code style update (formatting, local variables)
[ ] Refactoring (no functional changes, no api changes)
[ ] Build related changes
[x] CI related changes
[ ] Infrastructure changes
[ ] Other... Please describe:
```

## What is the current behavior?
CircleCI is not enabled to create snapshots, add tags, or update the docs on commits to `master`.
<!-- Please describe the current behavior that you are modifying and link to one or more relevant issues. -->
Issue Number: 
Relates to #11592

## What is the new behavior?
After a commit to master, CircleCI runs lint and tests against different AngularJS versions. If those are successful and the commit was to the `master` branch, then it snapshots, commits to `angular/code.material.angularjs.org`, and deploys the updated docs site. It also commits and tags the code snapshot in `angular/bower-material`.

## Does this PR introduce a breaking change?
```
[ ] Yes
[x] No
```
<!-- If this PR contains a breaking change, please describe the impact and migration path for existing applications below. -->
<!-- Note that breaking changes are highly unlikely to get merged to master unless the validation is clear and the use case is critical. -->

## Other information
The list of tasks related to this and their order can be found in https://github.com/angular/material/issues/11592.

**Caretaker note:** This can't be merged until the GitHub tokens get set up in the CircleCI environment variables as described in #11592.